+++ src/MSALWrapper/Constants.cs
@@ -23,6 +23,11 @@ namespace Microsoft.Authentication.MSALWrapper
         public static readonly Uri AadRedirectUri = new Uri("http://localhost");
 
         /// <summary>
+        /// The redirect uri for macOS brokered authentication (unsigned runtime apps).
+        /// </summary>
+        public static readonly Uri MacOSBrokerRedirectUri = new Uri("msauth.com.msauth.unsignedapp://auth");
+
+        /// <summary>
         /// The name of an environment variable used to disable file cache configuration.
         /// </summary>
         internal const string OEAUTH_MSAL_DISABLE_CACHE = "OEAUTH_MSAL_DISABLE_CACHE";